Refactor peer CLIs to a separate cli folder
