+++ packages/peer/package.json
@@ -22,8 +22,8 @@
     "build": "tsc",
     "lint": "eslint .",
     "dev": "node dist/index.js",
-    "create-peer": "node dist/create-peer.js",
-    "relay-node": "node dist/relay.js"
+    "create-peer": "node dist/cli/create-peer.js",
+    "relay-node": "node dist/cli/relay.js"
   },
   "dependencies": {
     "@cerc-io/libp2p": "0.42.2-laconic-0.1.1",
