+++ packages/peer/src/relay.ts
@@ -4,10 +4,6 @@
 
 import { Libp2p, createLibp2p } from '@cerc-io/libp2p';
 import wrtc from 'wrtc';
-import { hideBin } from 'yargs/helpers';
-import yargs from 'yargs';
-import fs from 'fs';
-import path from 'path';
 import debug from 'debug';
 
 import { noise } from '@chainsafe/libp2p-noise';
@@ -15,7 +11,6 @@ import { mplex } from '@libp2p/mplex';
 import { WebRTCDirectNodeType, webRTCDirect } from '@cerc-io/webrtc-direct';
 import { floodsub } from '@libp2p/floodsub';
 import { pubsubPeerDiscovery } from '@libp2p/pubsub-peer-discovery';
-import { createFromJSON } from '@libp2p/peer-id-factory';
 import type { Connection } from '@libp2p/interface-connection';
 import { multiaddr } from '@multiformats/multiaddr';
 import type { PeerId } from '@libp2p/interface-peer-id';
@@ -26,53 +21,12 @@ import { dialWithRetry } from './utils/index.js';
 
 const log = debug('laconic:relay');
 
-const DEFAULT_HOST = '127.0.0.1';
-const DEFAULT_PORT = 9090;
-const DEFAULT_MAX_DIAL_RETRY = 5;
-
-interface Arguments {
-  host: string;
-  port: number;
-  announce?: string;
-  peerIdFile?: string;
-  relayPeers?: string;
-  maxDialRetry: number;
-}
-
-async function main (): Promise<void> {
-  const argv: Arguments = _getArgv();
-  let peerId: PeerId | undefined;
-  let relayPeersList: string[] = [];
-
-  if (argv.peerIdFile) {
-    const peerIdFilePath = path.resolve(argv.peerIdFile);
-    console.log(`Reading peer id from file ${peerIdFilePath}`);
-
-    const peerIdObj = fs.readFileSync(peerIdFilePath, 'utf-8');
-    const peerIdJson = JSON.parse(peerIdObj);
-    peerId = await createFromJSON(peerIdJson);
-  } else {
-    console.log('Creating a new peer id');
-  }
-
-  if (argv.relayPeers) {
-    const relayPeersFilePath = path.resolve(argv.relayPeers);
-
-    if (!fs.existsSync(relayPeersFilePath)) {
-      console.log(`File at given path ${relayPeersFilePath} not found, exiting`);
-      process.exit();
-    }
-
-    console.log(`Reading relay peer multiaddr(s) from file ${relayPeersFilePath}`);
-    const relayPeersListObj = fs.readFileSync(relayPeersFilePath, 'utf-8');
-    relayPeersList = JSON.parse(relayPeersListObj);
-  }
-
-  const listenMultiaddrs = [`/ip4/${argv.host}/tcp/${argv.port}/http/p2p-webrtc-direct`];
+export async function createRelayNode (listenHost: string, listenPort: number, relayPeersList: string[], maxDialRetry: number, announceDomain?: string, peerId?: PeerId): Promise<Libp2p> {
+  const listenMultiaddrs = [`/ip4/${listenHost}/tcp/${listenPort}/http/p2p-webrtc-direct`];
   const announceMultiaddrs = [];
 
-  if (argv.announce) {
-    announceMultiaddrs.push(`/dns4/${argv.announce}/tcp/443/https/p2p-webrtc-direct`);
+  if (announceDomain) {
+    announceMultiaddrs.push(`/dns4/${announceDomain}/tcp/443/https/p2p-webrtc-direct`);
   }
 
   const node = await createLibp2p({
@@ -154,7 +108,7 @@ async function main (): Promise<void> {
         remoteAddr,
         {
           redialDelay: RELAY_REDIAL_DELAY,
-          maxRetry: argv.maxDialRetry
+          maxRetry: maxDialRetry
         }
       ).catch((error: Error) => console.log(error.message));
     }
@@ -162,48 +116,10 @@ async function main (): Promise<void> {
 
   if (relayPeersList.length) {
     console.log('Dialling relay peers');
-    await _dialRelayPeers(node, relayPeersList, argv.maxDialRetry);
+    await _dialRelayPeers(node, relayPeersList, maxDialRetry);
   }
-}
 
-function _getArgv (): Arguments {
-  return yargs(hideBin(process.argv)).parserConfiguration({
-    'parse-numbers': false
-  }).options({
-    host: {
-      type: 'string',
-      alias: 'h',
-      default: DEFAULT_HOST,
-      describe: 'Host to bind to'
-    },
-    port: {
-      type: 'number',
-      alias: 'p',
-      default: DEFAULT_PORT,
-      describe: 'Port to start listening on'
-    },
-    announce: {
-      type: 'string',
-      alias: 'a',
-      describe: 'Domain name to be used in the announce address'
-    },
-    peerIdFile: {
-      type: 'string',
-      alias: 'f',
-      describe: 'Relay Peer Id file path (json)'
-    },
-    relayPeers: {
-      type: 'string',
-      alias: 'r',
-      describe: 'Relay peer multiaddr(s) list file path (json)'
-    },
-    maxDialRetry: {
-      type: 'number',
-      describe: 'Maximum number of retries for dialling a relay peer',
-      default: DEFAULT_MAX_DIAL_RETRY
-    }
-  // https://github.com/yargs/yargs/blob/main/docs/typescript.md?plain=1#L83
-  }).parseSync();
+  return node;
 }
 
 async function _dialRelayPeers (node: Libp2p, relayPeersList: string[], maxDialRetry: number): Promise<void> {
@@ -226,7 +142,3 @@ async function _handleDeadConnections (node: Libp2p, remotePeerId: PeerId): Prom
   await node.hangUp(remotePeerId);
   log('Closed');
 }
-
-main().catch(err => {
-  console.log(err);
-});
